feat: removes no longer working and "hacky" "Bookmarks"-card
The card did no longer work.
Furthermore it used a error-prone hack that cloned the bookmarks
from the administration sidebar.
Currently the only control for bookmarks is the entire panel in the
administration sidebar. This cannot be loaded into the card's content.
Adding the bookmarks manually is too complicated as it requires complex
logic which usually the sidebar panel handles.

Therefore the card is removed for now.