feat: updated count and registration logic for users and orders
In this commit, various PHP and Javascript files related to getting user and order counts were
refactored. The changes updated the logic for getting a count of users and registrations.
Adjustments were also made to the grid layout in the SaaS dashboard. Some aspects of this refactor
included:
- Changing 'getUsersCount' to 'getOrdersCount'
- Modifying the return type from 'int' to 'int|string' in some functions
- Updating method 'package_quiqqer_dashboard_ajax_backend_saas_getUsersCount' to
'package_quiqqer_dashboard_ajax_backend_saas_getOrdersCount'
- Adjusting column classes in the SaaS dashboard for better layout responsiveness
- Modifying locale strings
- Adjusting default behaviour for some dashboard cards
- Including 'OrdersCount' and 'UsersCount' cards in the SaaS dashboard
The overall aim of these changes was to improve the functionality and user interface of the backend
system.